sentiment_agent/requirements.txt
h11~=0.16.0
rsa~=4.9.1
pyasn1~=0.6.1
grpcio~=1.73.1
requests~=2.32.4
rich~=14.0.0
ipython~=9.4.0
notebook~=7.4.4
PyYAML~=6.0.2
typing_extensions~=4.14.1
sniffio~=1.3.1
click~=8.2.1
Pygments~=2.19.2
httpcore~=1.0.9
idna~=3.10
httpx~=0.28.1
zstandard~=0.23.0
protobuf~=5.29.5
proto-plus~=1.26.1
sympy~=1.14.0
mpmath~=1.3.0
numpy~=2.3.1
decorator~=5.2.1
torch~=2.7.1
matplotlib~=3.10.3
filelock~=3.18.0
setuptools~=80.9.0
tqdm~=4.67.1
psutil~=7.0.0
packaging~=24.2
networkx~=3.5
urllib3~=2.5.0
fsspec~=2025.5.1
aiohttp~=3.12.14
yarl~=1.20.1
pyasn1_modules~=0.4.2
cachetools~=5.5.2
certifi~=2025.7.9
cffi~=1.17.1
google-api-python-client~=2.176.0
grpcio-status~=1.71.2
pydantic~=2.11.7
pillow~=11.3.0
pandas~=2.3.1
httplib2~=0.22.0
MarkupSafe~=3.0.2
orjson~=3.10.18
xxhash~=3.5.0
anyio~=4.9.0
pyparsing~=3.2.3
pydantic_core~=2.33.2
annotated-types~=0.7.0
charset-normalizer~=3.4.2
tenacity~=9.1.2
tornado~=6.5.1
langgraph~=0.5.2
langchain-core~=0.3.68
ormsgpack~=1.10.0
langsmith~=0.4.4
openai~=1.95.0
requests-toolbelt~=1.0.0
langchain~=0.3.26
Jinja2~=3.1.6
uritemplate~=4.2.0
jsonpointer~=3.0.0
jsonpatch~=1.33
langchain-text-splitters~=0.3.8
transformers~=4.53.2
google-auth-httplib2~=0.2.0
python-dotenv~=1.1.1
Flask~=3.1.1